src/transifex/requirements.txt
Django==1.3.1
Markdown==2.0.3
PIL==1.1.7
Pygments==1.3.1
South==0.7.2
contact-form==0.3
django-addons==0.6.6
django-ajax-selects==1.1.4
django-authority==0.4
django-filter==0.5.3
django-notification==0.1.5
django-pagination==1.0.7
django-staticfiles==0.3.2
django-tagging==0.3.1
polib==0.6.3
pygooglechart==0.3.0
python-magic==0.3.1
django-tagging-autocomplete==0.3.1
redis==2.4.10

-f http://trac.transifex.org/files/deps/django-piston-0.2.3-devel-r278.tar.gz
-f http://trac.transifex.org/files/deps/django-sorting-0.1.tar.gz
-f http://trac.transifex.org/files/deps/django-threadedcomments-0.9.tar.gz
